Refactors node handle to model package
Moves the node handle type to the `model` package for better
organization and separation of concerns. This change ensures that
data structures are clearly defined within the appropriate package.
